c2c_django/requirements.txt
Django==2.0.7
django-tastypie==0.14.3
django-cors-headers==3.4.0